+++ src/imio/helpers/content.py
@@ -691,13 +691,14 @@ def object_ids(context, class_names):
     return res
 
 
-def get_user_fullname(userid, none_if_no_user=False, none_if_unfound=False):
+def get_user_fullname(userid, none_if_no_user=False, none_if_unfound=False, with_user_id=False):
     """Get fullname without using getMemberInfo that is slow slow slow...
     We get it only from mutable_properties or authentic.
 
     :param userid: principal id
     :param none_if_no_user: return None if principal is not a user
     :param none_if_unfound: return None if principal is not found
+    :param with_user_id: include user_id between () in the returned result
     :return: fullname or userid if fullname is empty.
     """
     userid = safe_unicode(userid)
@@ -734,7 +735,10 @@ def get_user_fullname(userid, none_if_no_user=False, none_if_unfound=False):
         info = mt.getMemberInfo(userid)
         if info:
             fullname = info.get('fullname', '')
-    return safe_unicode(fullname) or userid
+    fullname = safe_unicode(fullname) or userid
+    if with_user_id:
+        fullname = u'{0} ({1})'.format(fullname, userid)
+    return fullname
 
 
 def sort_on_vocab_order(values, vocab=None, obj=None, vocab_name=None):